requirements.txt
Pillow>=10.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
